Resolve logic errors from merge.
