--- src/main/java/seedu/duke/logic/commands/ListFavCommand.java
@@ -1,10 +1,0 @@
-package seedu.duke.logic.commands;
-
-import seedu.duke.exceptions.CustomException;
-import seedu.duke.logic.commands.commons.Command;
-
-public class ListFavCommand extends Command {
-    @Override
-    public void executeCommand() throws CustomException {
-    }
-}
